Move sync job to concerns
